movie_explorer_backend/requirement.txt
Django==5.2.4 
django-cors-headers==4.7.0 
django-filter==23.2  
djangorestframework==3.16.0 
drf-spectacular==0.28.0 
psycopg2-binary==2.9.10 
pytest==8.4.1
pytest-django==4.11.1